+++ packages/scripts/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @bike4mind/scripts
 
+## 0.1.65
+
+### Patch Changes
+
+- Updated dependencies [[`3a3aef0`](https://github.com/Bike4Mind/bike4mind/commit/3a3aef0b59afe349f2f5e78ff3c693ea98f616e7), [`da0f7ab`](https://github.com/Bike4Mind/bike4mind/commit/da0f7abd8decb98b13cb3c006a2a4f21e294a974), [`1a01f97`](https://github.com/Bike4Mind/bike4mind/commit/1a01f977b4b3cfff9a325ac2867553f4ee8b79e7), [`cae26c4`](https://github.com/Bike4Mind/bike4mind/commit/cae26c4026e606cf412aa9996fdb7a9b444a0e8f), [`8e03a0e`](https://github.com/Bike4Mind/bike4mind/commit/8e03a0ed6430e40280db316e2301a0f20a8ddc57), [`16a8920`](https://github.com/Bike4Mind/bike4mind/commit/16a8920bd5a64e14d5fdd73fe0eb14da309d7866), [`c9f2085`](https://github.com/Bike4Mind/bike4mind/commit/c9f208569698a2a1ec8210923493d1c460cefbca), [`0b4e580`](https://github.com/Bike4Mind/bike4mind/commit/0b4e58050f10e92ec4f6fad32017d28c54a9d0ae), [`9fad658`](https://github.com/Bike4Mind/bike4mind/commit/9fad658b6504fa00b85045b028aa23c8d27d7bb2), [`472f90d`](https://github.com/Bike4Mind/bike4mind/commit/472f90d7f9387a879757ffa81746845ad93a93b2), [`9fc991e`](https://github.com/Bike4Mind/bike4mind/commit/9fc991e214af4fd2b1442759dc37528e74b33f11), [`5139e4e`](https://github.com/Bike4Mind/bike4mind/commit/5139e4e712089049eb47519455232c5e219c0fbd), [`de702ea`](https://github.com/Bike4Mind/bike4mind/commit/de702ea4ada1f91ad26167f2d7899a336cf647da), [`50b52a5`](https://github.com/Bike4Mind/bike4mind/commit/50b52a5fb5f3344b56bd4644b3a2154ca51fe31e), [`cb58a53`](https://github.com/Bike4Mind/bike4mind/commit/cb58a5394c377c8250ff7dafa0a251652be2a87a), [`82cd12c`](https://github.com/Bike4Mind/bike4mind/commit/82cd12c90c7772735e2c00c79b0a028fc0c890bd), [`7c8240c`](https://github.com/Bike4Mind/bike4mind/commit/7c8240ce7aa7ab839ad3ac7cc42aa51bc4fa9055), [`c46c8a4`](https://github.com/Bike4Mind/bike4mind/commit/c46c8a46e33df208d4547be6cd07b79add171ef2), [`e805cbe`](https://github.com/Bike4Mind/bike4mind/commit/e805cbe54ebd5c7d1113769d9e28875b79a71fe9), [`1507c14`](https://github.com/Bike4Mind/bike4mind/commit/1507c143605a375cce15735d4a953c3ee470bc7d)]:
+  - @bike4mind/common@5.0.0
+  - @bike4mind/services@6.0.0
+  - @bike4mind/utils@4.1.0
+  - @bike4mind/fab-pipeline@1.1.0
+  - @bike4mind/llm-adapters@0.11.2
+  - @bike4mind/database@0.1.60
+
 ## 0.1.64
 
 ### Patch Changes